refactor: 무한스크롤 로직을 PostContainer컴포넌트로 이동
- 메인페이지에서 사용한 PostContainer를 프로필페이지에서 재사용하기 위해 로직 이동
